requirements.txt
azure-core==1.30.2
azure-identity==1.16.1
azure-keyvault-secrets==4.8.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
greenlet==3.0.3
idna==3.7
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
